azure_functions/azure_function_indexing/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions

# Azure Services
azure-ai-formrecognizer
azure-storage-blob
azure-search-documents
azure-identity

# LangChain and OpenAI
langchain
langchain-openai
langchain-community
langchain-core

# Utilities
python-dotenv